Fixed error in typedef struct MUTEX where it used uint32_t to stored mutex locked flag. Fixed error in kernel/hwdep/x86_64/Makefile.in where it used incorrect directory for -I flags. Changed return type for lock_mutex(), unlock_mutex(), initialize_mutex() from size_t to void. Updated code to initialize memory map in kernel/hwdep/x86_64/init.asm. Fixed error in kernel/hwdep/x86_64/Makefile.in where it used fixed arch string instead of @arch@. Fixed error in modules/drivers/pic/x86_64/picx86_64.asm where esp register was used instead or rsp. Fixed error in FAT12/FAT16 first stage loader and second stage loader where it loaded FAT sectors incorrectly. Fixed error in _asm_init() in kernel/hwdep/x86_64/init.asm where it did not calculate the address of the memory map correctly. Fixed error in devicemanager_int() where it did not set irq_handlers to NULL. Replaced constant UART clock rate number with UART_CLOCK_RATE define in serialio_init(). Fixed error in serialio_init() where it did not check if add_character_device() returned -1. Fixed error in serialio_init() where void function returned -1. Fixed error in serialio_init() where it did not return 0 at end. Fixed error in pic_init() where 0x8e instead of 0x8e00 was used to set IRQ interrupt handlers for x86_64 CPUs. Fixed error in setirqhandler() where it did not set end of list marker to NULL. Fixed error in irq() in picx86_64.asm where it pushed registers onto stack instead of loading into edi,esi. Added missing prototypes for string functions into header and included header into c files. Fixed error in serialio_init() and serialio_ioctl(), fat_update_free_block_count(),doline(),cd_command(),for_command(),exit_command(),dir_command(),setvar() where incorrect function parameter types and number of paramters were used, but were not caught because of no prototype. Removed unused prototype for tohex() in command.h. Fixed error in  getphysicaladdress() in kernel/hwdep/x86_64/page.c where it used the incorrect page tables to get the physical address. Moved pic.c to hwdep/<arch> and updated Makefile.in to reference new file location.
Added format and width specifiers to kprintf_direct():
	* Changed tohex() to not print leading zeroes
	* Added code for format and width to kprintf_direct()
	* Fixed error in kprintf_direct() where %x didn't print in lowercase
	* Added tolowercase() function
	* Updated exception() in kernel/hwdep/i386/ex.c, kernel/hwdep/x86_64/ex.c to use updated kprintf_direct() formats
	* Updated DEBUG_PRINT_HEX,DEBUG_PRINT_HEX_LONG macros

Replaced duplicate kprintf() code
	* Implemented stub outputconsole() function in command.c.
	* Updated command.c,copyfile.c,readline.c,runcommand.c
	* Updated command/Makefile.in

Fixed error in dir_command() where it did not pad out time and date digits.

Replaced duplicate getfullpath() code
	* Moved getfullpath() from kernel/vfs.c to kernel/getfullpath.c.
	* Deleted command/getfullpath.c
	* Updated kernel/Makefile.in,command/Makefile.in
	* Updated Makefile.in

Fixed error in kprintf_direct() where %o specifier did not print numbers in octal:
	* Implemented tooctal() function
	* Updated kprintf_direct() to call tooctal()

+++ header/mutex.h
@@ -2,7 +2,7 @@
 #include <stddef.h>
 
 typedef struct {
-	uint32_t is_locked;
+	size_t is_locked;
 	size_t owner_process;
 } MUTEX;
